feat: endpoint created

+++ src/main/java/grp/nfe/service/ClienteService.java
@@ -5,7 +5,6 @@ import grp.nfe.repository.ClienteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 
@@ -15,42 +14,43 @@ public class ClienteService {
     @Autowired
     private ClienteRepository clienteRepository;
 
-    public List<Cliente> listarTodos() {
+    public Iterable<Cliente> buscarTodosClientes() {
         return clienteRepository.findAll();
     }
 
     public Cliente buscarPorCodigo(Integer codigo) {
         Optional<Cliente> cliente = clienteRepository.findByCodigo(codigo);
-        return cliente.orElseThrow(() -> new NoSuchElementException("Cliente não encontrado: " + codigo));
+        return cliente.orElseThrow(() -> new NoSuchElementException("Nenhum cliente encontrado com o código fornecido"));
+    }
+
+    public Cliente buscarPorNome(String nome) {
+        Optional<Cliente> cliente = clienteRepository.findByNome(nome);
+        return cliente.orElseThrow(() -> new NoSuchElementException("Nenhum cliente encontrado com o nome fornecido!"));
     }
 
     public Cliente create(Cliente cliente) {
-        if (cliente.getCodigo() == null) {
-            throw new IllegalArgumentException("Código do cliente é obrigatório!");
-        }
-        if (cliente.getNome() == null || cliente.getNome() == "" || cliente.getNome().isBlank()) {
-            throw new IllegalArgumentException("Nome do cliente é obrigatório!");
-        }
-        if (clienteRepository.existByCodigo(cliente.getCodigo())) {
-            throw new IllegalArgumentException("Já existe um cliente cadastro com o código fornecido: " + cliente.getCodigo());
+        if (clienteRepository.findByCodigo(cliente.getCodigo()).isPresent()) {
+            throw new IllegalArgumentException("Já existe cliente cadastrado com o mesmo código!");
         }
         return clienteRepository.save(cliente);
     }
 
     public Cliente update(Cliente cliente) {
-        if (cliente.getNome() == null || cliente.getNome().isBlank() || cliente.getNome() == "") {
-            throw new IllegalArgumentException("Nome do cliente é obrigatório!");
+        if (!clienteRepository.findByCodigo(cliente.getCodigo()).isPresent()) {
+            throw new IllegalArgumentException("Não existe cliente cadastrado com este código!");
         }
-        Cliente clienteAntigo = buscarPorCodigo(cliente.getCodigo());
+        Cliente clienteAntigo = clienteRepository.findByCodigo(cliente.getCodigo()).get();
         clienteAntigo.setNome(cliente.getNome());
         return clienteRepository.save(clienteAntigo);
     }
 
     public void delete(Integer codigo) {
-        Cliente cliente =  buscarPorCodigo(codigo);
-        clienteRepository.delete(cliente);
+        if (codigo == null || !clienteRepository.existsByCodigo(codigo)) {
+            throw new IllegalArgumentException("Não foi possível encontrar o cliente com o código especificado para exclusão!");
+        } else {
+            Cliente cliente = clienteRepository.findByCodigo(codigo).orElse(null);
+            clienteRepository.delete(cliente);
+        }
     }
-
-
 
 }
